Allow users to specify the attribute name
This allows user to use a column in the database that is named
something other than `attribute_at`. On the surface,
the primary usage would be if a user wanted to use a date column,
they would follow the convention of using `attribute_on`,
then specify that name in the `time_for_a_boolean column.
It also allows you to simply name your column whatever you prefer.

We probably shouldn't force people to follow conventions,
their column names may be out of their control.